adding exception handling for duplicate students and invalid marks; update marks type to float

+++ Student.java
@@ -6,10 +6,10 @@ public class Student {
     private long prn;
     private String name;
     private Date dob;
-    private double marks;
+    private float marks;
 
     // declaring constructors
-    public Student(long prn, String name, Date dob, double marks) {
+    public Student(long prn, String name, Date dob, float marks) {
         this.prn = prn;
         this.name = name;
         this.dob = dob;
@@ -45,7 +45,7 @@ public class Student {
         return marks;
     }
 
-    public void setMarks(double marks) {
+    public void setMarks(float marks) {
         this.marks = marks;
     }
